+++ decoder_test.go
@@ -38,7 +38,7 @@ func TestStringLiteralDecode(t *testing.T) {
 func TestIndexedLiteralDecode(t *testing.T) {
 	// D.2.2.  Literal Header Field without Indexing
 	var indexing Indexing = WITHOUT
-	var index uint64 = 4
+	var index uint32 = 4
 	var value string = "/sample/path"
 	buf := []byte{
 		0x04, 0x0c,
@@ -61,7 +61,7 @@ func TestIndexedLiteralDecode(t *testing.T) {
 
 func TestIndexedHeaderDecode(t *testing.T) {
 	// D.2.4.  Indexed Header Field
-	var index uint64 = 2
+	var index uint32 = 2
 	buf := []byte{0x82}
 
 	sw := swrap.New(buf)
