+++ header_field_test.go
@@ -7,6 +7,6 @@ import (
 
 func TestHeaderFieldSize(t *testing.T) {
 	h := NewHeaderField("hello", "world")
-	var actual uint64 = 42
+	var actual uint32 = 42
 	assert.Equal(t, h.Size(), actual)
 }
