+++ type_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestNewIndexedHeader(t *testing.T) {
-	var index uint64 = 10
+	var index uint32 = 10
 	var frame *IndexedHeader
 	frame = NewIndexedHeader(index)
 
@@ -16,13 +16,13 @@ func TestNewIndexedHeader(t *testing.T) {
 
 func TestNewIndexedLiteral(t *testing.T) {
 	var indexing Indexing = WITH
-	var index uint64 = 10
+	var index uint32 = 10
 	var value string = "var"
 	var frame *IndexedLiteral = NewIndexedLiteral(indexing, index, value)
 
 	assert.Equal(t, frame.Indexing, indexing)
 	assert.Equal(t, frame.Index, index)
-	assert.Equal(t, frame.ValueLength, uint64(len(value)))
+	assert.Equal(t, frame.ValueLength, uint32(len(value)))
 	assert.Equal(t, frame.ValueString, value)
 }
 
@@ -33,8 +33,8 @@ func TestNewStringLiteral(t *testing.T) {
 	var frame *StringLiteral = NewStringLiteral(indexing, name, value)
 
 	assert.Equal(t, frame.Indexing, indexing)
-	assert.Equal(t, frame.NameLength, uint64(len(name)))
+	assert.Equal(t, frame.NameLength, uint32(len(name)))
 	assert.Equal(t, frame.NameString, name)
-	assert.Equal(t, frame.ValueLength, uint64(len(value)))
+	assert.Equal(t, frame.ValueLength, uint32(len(value)))
 	assert.Equal(t, frame.ValueString, value)
 }
